move some fixtures into conftest
